requirements.txt
# Usage: pip install -r requirements.txt

PyYAML>=5.3.1
yt-dlp
matplotlib>=3.2.2
opencv-python>=4.1.1
Pillow>=7.1.2
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
protobuf<4.21.3
pandas>=1.1.4
seaborn>=0.11.0
numpy<1.24.0,>=1.18.5
flask
flask_cors
flask_socketio
ultralytics
python-telegram-bot
simple-websocket>=1.0.0

